requirements.txt
dataclasses==0.7
elevate==0.1.3
pkg-resources==0.0.0
pyusb==1.0.2
